解决trunk代码告警

+++ services/media_engine/plugins/source/http_source/download/downloader.cpp
@@ -314,8 +314,8 @@ bool Downloader::Seek(int64_t offset)
     if (offset >= 0 && offset < static_cast<int64_t>(contentLength)) {
         currentRequest_->startPos_ = offset;
     }
-    int64_t temp = currentRequest_->GetFileContentLength() - currentRequest_->startPos_;
-    currentRequest_->requestSize_ = static_cast<int>(std::min(temp, static_cast<int64_t>(PER_REQUEST_SIZE)));
+    size_t temp = currentRequest_->GetFileContentLength() - static_cast<size_t>(currentRequest_->startPos_);
+    currentRequest_->requestSize_ = static_cast<int>(std::min(temp, static_cast<size_t>(PER_REQUEST_SIZE)));
     currentRequest_->isEos_ = false;
     shouldStartNextRequest = false; // Reuse last request when seek
     return true;
